// serviceLayer/src/fbs/generated/model/patronV7.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * FBS Adapter
 * OpenAPI spec version: 1.0
 */
import type { AddressV2 } from "./addressV2";
import type { BlockStatus } from "./blockStatus";
import type { InterestV1 } from "./interestV1";
import type { Period } from "./period";

export interface PatronV7 {
  /** Tags associated with the patron. Can be used to e.g. give permissions to patron in external ressources. */
  tags?: string[];
  address?: AddressV2;
  /** True if the user is allowed to create bookings. */
  allowBookings?: boolean;
  birthday?: string;
  /** A list of block statuses -
 if the patron is not blocked then this value is empty or null */
  blockStatus?: BlockStatus[];
  /** Length of default interest period in days */
  defaultInterestPeriod: number;
  emailAddress?: string;
  guardianVisibility: boolean;
  interests?: InterestV1[];
  name?: string;
  /** Notification protocols that the patron want to receive notification on. SMS and EMAIL are not included. */
  notificationProtocols?: string[];
  onHold?: Period;
  /** Patron identifier to be used in subsequent service calls involving the patron */
  patronId: number;
  phoneNumber?: string;
  /** Language in which the patron prefers the communication with the library to take place */
  preferredLanguage?: string;
  /** ISIL of preferred pickup branch */
  preferredPickupBranch: string;
  receiveEmail: boolean;
  receivePostalMail: boolean;
  receiveSms: boolean;
  /** True if the user is resident in the same municipality as the library */
  resident: boolean;
  secondaryAddress?: AddressV2;
}